Unison.Codebase.Execute

Description

Execute a computation of type '{IO} () that has been previously added to the codebase, without setting up an interactive environment.

This allows one to run standalone applications implemented in the Unison language.

Documentation

execute :: Codebase IO Symbol Ann -> Runtime Symbol -> ProjectPathNames -> IO (Either Error ())

codebaseToCodeLookup :: MonadIO m => Codebase m Symbol Ann -> CodeLookup Symbol m Ann