Unison.Runtime.ANF

Documentation

minimizeCyclesOrCrash :: (Var v, Ord a) => Term v a -> Term v a

pattern TVar :: Var v => v -> ANormal ref v

pattern TLit :: Var v => Lit ref -> ANormal ref v

pattern TBLit :: Var v => Lit ref -> ANormal ref v

pattern TApp :: Var v => Func ref v -> [v] -> ANormal ref v

pattern TApv :: Var v => v -> [v] -> ANormal ref v

pattern TCom :: Var v => ref -> [v] -> ANormal ref v

pattern TCon :: Var v => ref -> CTag -> [v] -> ANormal ref v

pattern UFalse :: Var v => ANormal TypeReference v

pattern UTrue :: Var v => ANormal TypeReference v

pattern TKon :: Var v => v -> [v] -> ANormal ref v

pattern TReq :: Var v => ref -> CTag -> [v] -> ANormal ref v

pattern TPrm :: Var v => POp -> [v] -> ANormal ref v

pattern TFOp :: Var v => ForeignFunc -> [v] -> ANormal ref v

pattern THnd :: Var v => [ref] -> v -> Maybe v -> ANormal ref v -> ANormal ref v

pattern TLet :: Var v => Direction Word16 -> v -> Mem -> ANormal ref v -> ANormal ref v -> ANormal ref v

pattern TLetD :: Var v => v -> Mem -> ANormal ref v -> ANormal ref v -> ANormal ref v

pattern TFrc :: Var v => v -> ANormal ref v

pattern TLets :: Var v => Direction Word16 -> [v] -> [Mem] -> ANormal ref v -> ANormal ref v -> ANormal ref v

pattern TName :: Var v => v -> Either ref v -> [v] -> ANormal ref v -> ANormal ref v

pattern TBind :: Var v => Cte v -> ANormal Reference v -> ANormal Reference v

pattern TBinds :: Var v => [Cte v] -> ANormal Reference v -> ANormal Reference v

pattern TShift :: Var v => ref -> v -> ANormal ref v -> ANormal ref v

pattern TMatch :: Var v => v -> Branched ref (ANormal ref v) -> ANormal ref v

pattern TDiscard :: Var v => v -> ANormal ref v

pattern TLocal :: Var v => v -> ANormal ref v -> ANormal ref v

pattern TUpdate :: Var v => Bool -> v -> v -> Term (ANormalF ref) v

data FloatName v

Constructors

FloatName [FloatSeg v]

prettyFloatName :: Var v => PrettyPrintEnv -> FloatName v -> Pretty ColorText

data Mem

Constructors

UN

BX

data Lit ref

Constructors

I Int64

N Word64

F Double

T Text

C Char

LM (Referent' ref)

LY ref

data Cacheability

Whether the evaluation of a given definition is cacheable or not. i.e. it's a top-level pure value.

Constructors

Cacheable

Uncacheable

data Direction a

Constructors

Indirect a

Direct

data SuperNormal ref v

Constructors

Lambda

Fields

conventions :: [Mem]

bound :: ANormal ref v

arity :: SuperNormal ref v -> Int

data SuperGroup ref v

Constructors

Rec

Fields

group :: [(v, SuperNormal ref v)]

entry :: SuperNormal ref v

arities :: SuperGroup ref v -> [Int]

data POp

Constructors

ADDI

SUBI

MULI

DIVI

SGNI

NEGI

MODI

POWI

SHLI

SHRI

ANDI

IORI

XORI

COMI

INCI

DECI

LEQI

LESI

EQLI

NEQI

TRNC

ADDN

SUBN

DRPN

MULN

DIVN

MODN

TZRO

LZRO

POPC

POWN

SHLN

SHRN

ANDN

IORN

XORN

COMN

INCN

DECN

LEQN

LESN

EQLN

NEQN

ADDF

SUBF

MULF

DIVF

MINF

MAXF

LEQF

LESF

EQLF

NEQF

POWF

EXPF

SQRT

LOGF

LOGB

ABSF

CEIL

FLOR

TRNF

RNDF

COSF

ACOS

COSH

ACSH

SINF

ASIN

SINH

ASNH

TANF

ATAN

TANH

ATNH

ATN2

CATT

TAKT

DRPT

SIZT

IXOT

UCNS

USNC

EQLT

LEQT

PAKT

UPKT

CATS

TAKS

DRPS

SIZS

CONS

SNOC

IDXS

BLDS

VWLS

VWRS

SPLL

SPLR

PAKB

UPKB

TAKB

DRPB

IXOB

IDXB

SIZB

FLTB

CATB

ITOF

NTOF

ITOT

NTOT

TTOI

TTON

TTOF

FTOT

CAST

FORK

EQLU

CMPU

LEQU

LESU

EROR

MISS

CACH

LKUP

LOAD

CVLD

SDBX

VALU

TLTT

PRNT

INFO

TRCE

DBTX

ATOM

TFRC

SDBL

SDBV

REFN

REFR

REFW

RCAS

RRFC

TIKR

NOTB

ANDB

IORB

KEEP

FGNN

FGNF

close :: (Var v, Monoid a) => Set v -> Term v a -> Term v a

saturate :: (Var v, Monoid a) => Map ConstructorReference Int -> Term v a -> Term v a

float :: (Var v, Monoid a) => Map v Reference -> Term v a -> (Term v a, Map Reference Reference, Map Reference (FloatName v), [(Reference, Term v a)], [(Reference, Term v a)])

floatGroup :: (Var v, Monoid a) => Map v Reference -> [(v, Term v a)] -> ([(v, Id)], [(Reference, FloatName v)], [(Reference, Term v a)], [(Reference, Term v a)])

lamLift :: (Var v, Monoid a) => Map v Reference -> Term v a -> (Term v a, Map Reference Reference, Map Reference (FloatName v), [(Reference, Term v a)], [(Reference, Term v a)])

lamLiftGroup :: (Var v, Monoid a) => Map v Reference -> [(v, Term v a)] -> ([(v, Id)], [(Reference, FloatName v)], [(Reference, Term v a)], [(Reference, Term v a)])

litRef :: Lit ref -> Reference

inlineAlias :: (Var v, Monoid a) => Term v a -> Term v a

addDefaultCases :: (Var v, Monoid a) => Text -> Term v a -> Term v a

data ANormalF ref v e

Constructors

ALet (Direction Word16) [Mem] e e

AName (Either ref v) [v] e

ALit (Lit ref)

ABLit (Lit ref)

AMatch v (Branched ref e)

AShift ref e

AHnd [ref] v (Maybe v) e

AApp (Func ref v) [v]

AFrc v

AVar v

ADiscard v

ALocal v e

AUpdate Bool v v

Bundled Patterns

pattern AApv :: v -> [v] -> ANormalF ref v e

pattern ACom :: ref -> [v] -> ANormalF ref v e

pattern ACon :: ref -> CTag -> [v] -> ANormalF ref v e

pattern AKon :: v -> [v] -> ANormalF ref v e

pattern AReq :: ref -> CTag -> [v] -> ANormalF ref v e

pattern APrm :: POp -> [v] -> ANormalF ref v e

pattern AFOp :: ForeignFunc -> [v] -> ANormalF ref v e

type ANormal ref = Term (ANormalF ref)

data RTag

data CTag

newtype PackedTag

A combined tag, which is a packed representation of an RTag and a CTag

Constructors

PackedTag Word64

class Tag t where

Methods

rawTag :: t -> Word64

data GroupRef ref

Constructors

GR ref Word64

data Code ref

Constructors

CodeRep (SuperGroup ref Symbol) Cacheability

type ValList ref = [Value ref]

A list of either unboxed or boxed values. Each slot is one of unboxed or boxed but not both.

data Value ref

Constructors

Partial (GroupRef ref) (ValList ref)

Data ref Word64 (ValList ref)

Cont (ValList ref) (Cont ref)

BLit (BLit ref)

data Cont ref

Constructors

KE

Mark Word64 [ref] [(ref, Value ref)] (Cont ref)

Push Word64 Word64 (GroupRef ref) (Cont ref)

data BLit ref

Constructors

Text Text

List (Seq (Value ref))

TmLink (Referent' ref)

TyLink ref

Bytes Bytes

Quote (Value ref)

Code (Code ref)

BArr ByteArray

Arr (Array (Value ref))

Pos Word64

Neg Word64

Char Char

Float Double

Map [(Value ref, Value ref)]

BigInt Integer

BigNat Natural

packTags :: RTag -> CTag -> PackedTag

unpackTags :: PackedTag -> (RTag, CTag)

maskTags :: PackedTag -> Word64

type ANFM v = ReaderT (Set v) (State (Word64, Word16, [(v, SuperNormal Reference v)]))

data Branched ref e

Constructors

MatchIntegral (EnumMap Word64 e) (Maybe e)

MatchText (Map Text e) (Maybe e)

MatchBytes (Map Bytes e) (Maybe e)

MatchRequest [(ref, EnumMap CTag ([Mem], e))] e

MatchEmpty

MatchData ref (EnumMap CTag ([Mem], e)) (Maybe e)

MatchSum (EnumMap Word64 ([Mem], e))

MatchNumeric ref (EnumMap Word64 e) (Maybe e)

Bundled Patterns

pattern MatchDataCover :: ref -> EnumMap CTag ([Mem], e) -> Branched ref e

data Func ref v

Constructors

FVar v

FComb !ref

FCont v

FCon !ref !CTag

FReq !ref !CTag

FPrim (Either POp ForeignFunc)

data SGEqv ref v

Constructors

NumDefns (SuperGroup ref v) (SuperGroup ref v)

DefnConventions (SuperNormal ref v) (SuperNormal ref v)

Subterms (ANormal ref v) (ANormal ref v)

equivocate :: (Ord ref, Var v) => SuperGroup ref v -> SuperGroup ref v -> Either (SGEqv ref v) ()

superNormalize :: Var v => Term v a -> SuperGroup Reference v

anfTerm :: Var v => Term v a -> ANFM v (DNormal v)

codeGroup :: Code ref -> SuperGroup ref Symbol

valueTermLinks :: Ord ref => Value ref -> [ref]

collectValueLinks :: Ord ref => Value ref -> (Set ref, Set ref)

valueLinks :: Monoid a => (Bool -> ref -> a) -> Value ref -> a

groupTermLinks :: (Ord ref, Var v) => SuperGroup ref v -> [ref]

replaceConstructors :: (Ord ref, Var v) => Map ref (Map CTag ForeignFunc) -> SuperGroup ref v -> SuperGroup ref v

replaceFunctions :: (Ord ref, Var v) => Map ref ref -> SuperGroup ref v -> SuperGroup ref v

foldGroup :: Monoid m => (SuperGroup ref Symbol -> m) -> Code ref -> m

foldGroupLinks :: (Monoid r, Var v) => (Bool -> ref -> r) -> SuperGroup ref v -> r

overGroup :: (SuperGroup ref0 Symbol -> SuperGroup ref1 Symbol) -> Code ref0 -> Code ref1

overGroupLinks :: Var v => (Bool -> ref0 -> ref1) -> SuperGroup ref0 v -> SuperGroup ref1 v

traverseGroup :: Applicative f => (SuperGroup ref0 Symbol -> f (SuperGroup ref1 Symbol)) -> Code ref0 -> f (Code ref1)

traverseGroupLinks :: (Applicative f, Var v) => (Bool -> ref0 -> f ref1) -> SuperGroup ref0 v -> f (SuperGroup ref1 v)

normalLinks :: (Applicative f, Var v) => (Bool -> ref0 -> f ref1) -> SuperNormal ref0 v -> f (SuperNormal ref1 v)

prettyGroup :: Var v => String -> SuperGroup Reference v -> ShowS

prettySuperNormal :: Var v => Int -> SuperNormal Reference v -> ShowS

prettyANF :: Var v => Bool -> Int -> ANormal Reference v -> ShowS