Unison.Runtime.ANF.Serialize

Documentation

data Version

Constructors

Transfer Word32

Hash Word32

index :: Eq v => [v] -> v -> Maybe Word64

deindex :: HasCallStack => [v] -> Word64 -> v

pushCtx :: [v] -> [v] -> [v]

putIndex :: Word64 -> Builder

getIndex :: forall (m :: Type -> Type). PrimBase m => Get m Word64

putVar :: Eq v => [v] -> v -> Builder

getVar :: forall (m :: Type -> Type) v. PrimBase m => [v] -> Get m v

putArgs :: Eq v => [v] -> [v] -> Builder

getArgs :: forall (m :: Type -> Type) v. PrimBase m => [v] -> Get m [v]

putCCs :: [Mem] -> Builder

getCCs :: forall (m :: Type -> Type). PrimBase m => Get m [Mem]

putGroup :: Var v => Map Reference Word64 -> Bool -> SuperGroup Reference v -> Builder

getGroup :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => GDeserial m (SuperGroup Reference v)

putCode :: Bool -> Code Reference -> Builder

getCode :: forall (m :: Type -> Type). PrimBase m => GDeserial m (Code Reference)

putInlineInfo :: Var v => [v] -> InlineInfo Reference v -> Builder

getInlineInfo :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => [v] -> Word64 -> GDeserial m (InlineInfo Reference v)

putInlineExpr :: Var v => [v] -> ANormal Reference v -> Builder

getInlineExpr :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => [v] -> Word64 -> GDeserial m (ANormal Reference v)

putOptInfos :: Var v => OptInfos Reference v -> Builder

getOptInfos :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => Get m (OptInfos Reference v)

putInlineClass :: InlineClass -> Builder

getInlineClass :: forall (m :: Type -> Type). PrimBase m => Get m InlineClass

putCacheability :: Cacheability -> Builder

getCacheability :: forall (m :: Type -> Type). PrimBase m => GDeserial m Cacheability

putComb :: Var v => Map Reference Word64 -> Bool -> [v] -> SuperNormal Reference v -> Builder

getFresh :: Var v => Word64 -> v

getComb :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => [v] -> Word64 -> GDeserial m (SuperNormal Reference v)

putNormal :: Var v => Map Reference Word64 -> Bool -> [v] -> ANormal Reference v -> Builder

getNormal :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => [v] -> Word64 -> GDeserial m (ANormal Reference v)

putFunc :: Var v => Map Reference Word64 -> Bool -> [v] -> Func Reference v -> Builder

getFunc :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => [v] -> GDeserial m (Func Reference v)

putFOp :: ForeignFunc -> Builder

getFOp :: forall (m :: Type -> Type). PrimBase m => Get m ForeignFunc

putPOp :: POp -> Builder

getPOp :: forall (m :: Type -> Type). PrimBase m => Get m POp

putLit :: Lit Reference -> Builder

getLit :: forall (m :: Type -> Type). PrimBase m => Get m (Lit Reference)

putBLit :: Version -> BLit Reference -> Builder

putAsMap :: Version -> [(Value Reference, Value Reference)] -> Builder

getBLit :: forall (m :: Type -> Type). PrimBase m => GDeserial m (BLit Reference)

putRefs :: [Reference] -> Builder

getRefs :: forall (m :: Type -> Type). PrimBase m => Get m [Reference]

putBranches :: Var v => Map Reference Word64 -> Bool -> [v] -> Branched Reference (ANormal Reference v) -> Builder

getBranches :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => [v] -> Word64 -> GDeserial m (Branched Reference (ANormal Reference v))

putCase :: Var v => Map Reference Word64 -> Bool -> [v] -> ([Mem], ANormal Reference v) -> Builder

getCase :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => [v] -> Word64 -> GDeserial m ([Mem], ANormal Reference v)

putCTag :: CTag -> Builder

getCTag :: forall (m :: Type -> Type). PrimBase m => Get m CTag

putGroupRef :: GroupRef Reference -> Builder

getGroupRef :: forall (m :: Type -> Type). PrimBase m => Get m (GroupRef Reference)

putValue :: Version -> Value Reference -> Builder

getValue :: forall (m :: Type -> Type). PrimBase m => GDeserial m (Value Reference)

putCont :: Version -> Cont Reference -> Builder

getCont :: forall (m :: Type -> Type). PrimBase m => GDeserial m (Cont Reference)

deserializeCode :: ByteString -> IO (Either String (Referenced Code))

serializeCode :: Bool -> Referenced Code -> ByteString

serializeCodeWithVersion :: Word64 -> Bool -> Referenced Code -> IO (Either String ByteString)

serializeGroupForRehash :: Var v => Reference -> SuperGroup Reference v -> ByteString

Serializes a SuperGroup for rehashing.

Expected as arguments are some code, and the Reference that refers to it. In particular, if the code refers to itself by reference, or if the code is part of a mututally-recursive set of definitions (which have a common hash), the reference used as part of that (mutual) recursion must be supplied.

Using that reference, we find all references in the code to that connected component. In the resulting byte string, those references are instead replaced by positions in a listing of the connected component. This means that the byte string is independent of the hash used for the self reference. Only the order matters (which is determined by the Reference). Then the bytes can be re-hashed to establish a new hash for the connected component. This operation should be idempotent as long as the indexing is preserved.

Supplying a Builtin reference is not supported. Such code shouldn't be subject to rehashing.

getVersionedValue :: forall (m :: Type -> Type). PrimBase m => Get m (Referenced Value)

deserializeValue :: ByteString -> IO (Either String (Referenced Value))

serializeValue :: Referenced Value -> ByteString

serializeValueWithVersion :: Word64 -> Referenced Value -> IO ByteString

serializeValueForHash :: Value Reference -> ByteString

getGroupCurrent :: forall (m :: Type -> Type) v. (PrimBase m, Var v) => Get m (SuperGroup Reference v)

type GDeserial (m :: Type -> Type) a = (Version, Bool) -> Get m a

type DeserialIO a = (Version, Bool) -> Get IO a

type DeserialST s a = (Version, Bool) -> Get (ST s) a

valueToCode :: Version -> Version

valueVersion :: Word32

codeVersion :: Word32