Unison.Runtime.Interface

Documentation

startRuntime :: Bool -> RuntimeHost -> Text -> IO (Runtime Symbol)

withRuntime :: MonadUnliftIO m => Bool -> RuntimeHost -> Text -> (Runtime Symbol -> m a) -> m a

standalone :: CCache () -> Word64 -> IO StoredCache

runStandalone :: Bool -> StoredCache -> CombIx -> IO (Either Error ())

data StoredCache

A version of the Code Cache designed to be serialized to disk as standalone bytecode.

Constructors

SCache (EnumMap Word64 Combs) (EnumMap Word64 Reference) (EnumSet Word64) (OptInfos Reference Symbol) (EnumMap Word64 Reference) Word64 Word64 (Map Reference (SuperGroup Reference Symbol)) (Map Reference Word64) (Map Reference Word64) (Map Reference (Set Reference))

decodeStandalone :: ByteString -> IO (Either String (Text, Text, CombIx, StoredCache))

data RuntimeHost

Whether the runtime is hosted within a persistent session or as a one-off process. This affects the amount of clean-up and book-keeping the runtime does.

Constructors

OneOff

Persistent

type Runtime = Runtime Error DecompError

terminate :: Runtime e e' v -> IO ()

Exported for tests

getStoredCache :: forall (m :: Type -> Type). PrimBase m => Get m StoredCache

putStoredCache :: StoredCache -> Builder

error serialization (these should live in unison-cli, but they still have some call sites here)

prettyError

Arguments

:: Applicative f

=> (Word -> f (Pretty ColorText))
A function for displaying unisonweb/unison issue numbers (for example, showIssueUrl).

-> Error

-> f (Pretty ColorText)

NB: The only reason this is in the unison-runtime package is because it’s used in the tests. Otherwise it should move to unison-cli.

prettyRuntimeExn :: Applicative f => (Word -> f (Pretty ColorText)) -> RuntimeExn -> f (Pretty ColorText)

renderDecompError :: DecompError -> Pretty ColorText

tabulateErrors :: Set DecompError -> Pretty ColorText