Unison.Runtime.Machine

Documentation

type ActiveThreads = Maybe (IORef (Set ThreadId))

A ref storing every currently active thread. This is helpful for cleaning up orphaned threads when the main process completes.

We track threads when running in a host process like UCM, otherwise, in one-off environments Nothing is used and we don't bother tracking forked threads since they'll be cleaned up automatically on process termination.

data CCache prof

Constructors

CCache

Fields

sandboxed :: Bool

tracer :: Bool -> Val -> Tracer

profiler :: !prof

srcCombs :: TVar (EnumMap Word64 Combs)

combs :: TVar (EnumMap Word64 MCombs)

combRefs :: TVar (EnumMap Word64 Reference)

cacheableCombs :: TVar (EnumSet Word64)

optInfos :: TVar (OptInfos Reference Symbol)

tagRefs :: TVar (EnumMap Word64 Reference)

freshTm :: TVar Word64

freshTy :: TVar Word64

intermed :: TVar (Map Reference (SuperGroup Reference Symbol))

refTm :: TVar (Map Reference Word64)

refTy :: TVar (Map Reference Word64)

sandbox :: TVar (Map Reference (Set Reference))

type Combs = GCombs Void CombIx

data Tracer

Constructors

NoTrace

MsgTrace String String String

SimpleTrace String

apply0 :: RuntimeProfiler p => Maybe (XStack -> IO ()) -> CCache p -> ActiveThreads -> Word64 -> IO ()

baseCCache :: Bool -> IO (CCache ())

cacheAdd :: RuntimeProfiler p => [(Reference, Code Reference)] -> CCache p -> IO [Reference]

cacheAdd0 :: RuntimeProfiler p => Set Reference -> [(Reference, Code Reference)] -> [(Reference, Set Reference)] -> CCache p -> IO ()

eval0 :: RuntimeProfiler p => CCache p -> ActiveThreads -> MSection -> IO ()

expandSandbox :: Map Reference (Set Reference) -> [(Reference, SuperGroup Reference Symbol)] -> [(Reference, Set Reference)]

preEvalTopLevelConstants :: RuntimeProfiler p => EnumMap Word64 (GCombs Val CombIx) -> EnumMap Word64 (GCombs Val CombIx) -> CCache p -> IO ()

refLookup :: String -> Map Reference Word64 -> Reference -> Word64

refNumTm :: CCache prof -> Reference -> IO Word64

refNumsTm :: CCache prof -> IO (Map Reference Word64)

refNumsTy :: CCache prof -> IO (Map Reference Word64)

reifyValue :: CCache p -> Referenced Value -> IO (Either [Reference] Val)

resolveSection :: CCache p -> Section -> IO MSection