Unison.Runtime.Serialize.Get

Documentation

newtype Get (m :: Type -> Type) a

Constructors

Get

Fields

unGet :: ByteString -> Ix m -> m a

data GetExn

Constructors

InsufficientBytes String

UserExn String

getExnMsg :: GetExn -> String

class PrimMonad m => PrimBase (m :: Type -> Type) where

Class of primitive monads for state-transformer actions.

Unlike PrimMonad, this typeclass requires that the Monad be fully expressed as a state transformer, therefore disallowing other monad transformers on top of the base IO or ST.

Since: primitive-0.6.0.0

Methods

internal :: m a -> State# (PrimState m) -> (# State# (PrimState m), a #)

Expose the internal structure of the monad.

newtype Ix (m :: Type -> Type)

Constructors

Ix (PrimVar (PrimState m) Int)

evaluated :: forall (m :: Type -> Type) a. PrimBase m => a -> Get m a

getByteString :: forall (m :: Type -> Type). PrimBase m => Int -> Get m ByteString

getWord8 :: forall (m :: Type -> Type). PrimBase m => Get m Word8

getVarInt :: forall int (m :: Type -> Type). (Bits int, Num int, Ord int, PrimBase m) => Get m int

getInt64be :: forall (m :: Type -> Type). PrimBase m => Get m Int64

getWord16be :: forall (m :: Type -> Type). PrimBase m => Get m Word16

getWord32be :: forall (m :: Type -> Type). PrimBase m => Get m Word32

getWord64be :: forall (m :: Type -> Type). PrimBase m => Get m Word64

getWord64le :: forall (m :: Type -> Type). PrimBase m => Get m Word64

getDoublebe :: forall (m :: Type -> Type). PrimBase m => Get m Double

getFloatbe :: forall (m :: Type -> Type). PrimBase m => Get m Float

getAccumulating :: forall (m :: Type -> Type) s a r. PrimBase m => s -> (s -> a -> s) -> (s -> r) -> Get m a -> Get m r

getAccumulatingRevList :: forall (m :: Type -> Type) a r. PrimBase m => ([a] -> r) -> Get m a -> Get m r

getArray :: forall (m :: Type -> Type) a. PrimBase m => Get m a -> Get m (Array a)

getList :: forall (m :: Type -> Type) a. PrimBase m => Get m a -> Get m [a]

getSeq :: forall (m :: Type -> Type) a. PrimBase m => Get m a -> Get m (Seq a)

getPrimArray :: forall (m :: Type -> Type) a. (PrimBase m, Prim a) => Get m a -> Get m (PrimArray a)

remaining :: forall (m :: Type -> Type). PrimBase m => Get m Int

runGet :: PrimBase m => Get m a -> ByteString -> m a

runGetCatch :: PrimBase m => Get m a -> ByteString -> m (Either String a)

runGetCatchIO :: Get IO a -> ByteString -> IO (Either String a)