Unison.Runtime.Stack

Documentation

data K

Constructors

KE

CB Callback

AMark !Int AEnv !AffineRef !K

Mark !Int !(EnumSet Word64) DEnv !K

Push !Int !Int !CombIx !Int !(RSection Val) !K

Local HEnv !Int !K

Keep !a !Int !K

data GClosure comb

Constructors

GPAp

Fields

!CombIx

!(GCombInfo comb)

!Seg

args

GEnum !Reference !PackedTag

GData1 !Reference !PackedTag !Val

GData2 !Reference !PackedTag !Val !Val

GDataG !Reference !PackedTag !Seg

GCaptured

Fields

!K

code cont

!Int

arg size

!Seg

u/b data stacks

GForeign !Foreign

GUnboxedTypeTag !UnboxedTypeTag
The type tag for the value in the corresponding unboxed stack slot.

We should consider adding separate constructors for common builtin type tags. GHC will optimize nullary constructors into singletons.

GAffine

Fields

!(EnumSet Word64)

associated ability numbers

!AEnv

original handler environment

!AffineRef

updateable reference

GBlackHole

newtype Closure

Constructors

Closure

Fields

unClosure :: GClosure (RComb Val)

Bundled Patterns

pattern DataC :: Reference -> PackedTag -> SegList -> Closure

pattern PApV :: CombIx -> RCombInfo Val -> SegList -> Closure

pattern CapV :: K -> Int -> SegList -> Closure

pattern PAp :: CombIx -> GCombInfo (RComb Val) -> Seg -> Closure

pattern Enum :: Reference -> PackedTag -> Closure

pattern Data1 :: Reference -> PackedTag -> Val -> Closure

pattern Data2 :: Reference -> PackedTag -> Val -> Val -> Closure

pattern DataG :: Reference -> PackedTag -> Seg -> Closure

pattern Captured :: K -> Int -> Seg -> Closure

pattern Foreign :: Foreign -> Closure

pattern Affine :: EnumSet Word64 -> AEnv -> AffineRef -> Closure

pattern BlackHole :: Closure

pattern UnboxedTypeTag :: UnboxedTypeTag -> Closure

data Failure a

Constructors

Failure Reference Text a

data Foreign

Constructors

WrapArray !(Array Val)

WrapByteArray !ByteArray

WrapBytes !Bytes

WrapCDynFunc !CDynFunc

WrapCPattern !CPattern

WrapCharPattern !CharPattern

WrapCode !(Referenced Code)

WrapClientSockAddr !ClientSockAddr

WrapDLL !DLL

WrapFFISpec !FFSpec

WrapFFIType !FFType

WrapHandle !Handle

WrapHashAlgorithm !HashAlgorithm

WrapIORef !(IORef Val)

WrapInteger !Integer

WrapListenSocket !ListenSocket

WrapMap !(Map Val Val)

WrapMutableByteArray !(MutableByteArray RealWorld)

WrapMutableArray !(MutableArray RealWorld Val)

WrapMVar !(MVar Val)

WrapNatural !Natural

WrapProcessHandle !ProcessHandle

WrapPromise !(Promise Val)

WrapPtr !(Ptr ())

WrapForeignPtr !(ForeignPtr ())

WrapReference !Reference

WrapReferent !Referent

WrapSeq !(Seq Val)

WrapSocket !Socket

WrapText !Text

WrapThreadId !ThreadId

WrapTicket !(Ticket Val)

WrapTimeSpec !TimeSpec

WrapTlsClientParams !ClientParams

WrapTlsServerParams !ServerParams

WrapTls !Tls

WrapTVar !(TVar Val)

WrapUDPSocket !UDPSocket

WrapValue !(Referenced Value)

WrapX509PrivKey !PrivKey

WrapX509SignedCertificate !SignedCertificate

foreignRef :: Foreign -> Reference

foreignName :: Foreign -> String

class BuiltinForeign f where

Methods

builtinName :: Tagged f String

wrapBuiltin :: f -> Foreign

maybeUnwrapBuiltin :: Foreign -> Maybe f

newtype AffineRef

Constructors

ARef (IORef Closure)

type AEnv = EnumMap Word64 AffineRef

type DEnv = EnumMap Word64 Val

data HEnv

Constructors

HEnv

Fields

aenv :: AEnv

denv :: DEnv

closureTag :: Closure -> PackedTag

formDataReplaced :: Reference -> PackedTag -> Seg -> Closure

unitClosure :: Closure

data UnboxedTypeTag

Constructors

CharTag

FloatTag

IntTag

NatTag

unboxedTypeTagToInt :: UnboxedTypeTag -> Int

unboxedTypeTagFromInt :: HasCallStack => Int -> UnboxedTypeTag

type IxClosure = GClosure CombIx

newtype Callback

Constructors

Hook (XStack -> IO ())

data Augment

Constructors

I

K

C

data Dump

Constructors

A

F Int Int

S

data Stack

Constructors

Stack

Fields

ap :: !Int

fp :: !Int

sp :: !Int

ustk :: !(MutableByteArray (PrimState IO))

bstk :: !(MutableArray (PrimState IO) Closure)

type XStack = (# Int#, Int#, Int#, MutableByteArray# (PrimState IO), MutableArray# (PrimState IO) Closure #)

pattern XStack :: Int# -> Int# -> Int# -> MutableByteArray# RealWorld -> MutableArray# RealWorld Closure -> Stack

packXStack :: XStack -> Stack

unpackXStack :: Stack -> XStack

xStackIOToIO :: IOXStack -> IO Stack

stackIOToIOX :: IO Stack -> IOXStack

estackIOToIOX :: IO (Bool, Stack) -> IOEXStack

exStackIOToIO :: IOEXStack -> IO (Bool, Stack)

type IOXStack = State# RealWorld -> (# State# RealWorld, XStack #)

type IOEXStack = State# RealWorld -> (# State# RealWorld, Bool, XStack #)

apX :: Stack -> Int#

fpX :: Stack -> Int#

spX :: Stack -> Int#

ustkX :: Stack -> MutableByteArray# RealWorld

bstkX :: Stack -> MutableArray# RealWorld Closure

type Off = Int

type SZ = Int

type FP = Int

type Seg = (USeg, BSeg)

type USeg = ByteArray

type BSeg = Array Closure

type SegList = [Val]

data Val

A runtime value, which is either a boxed or unboxed value, but we may not know which.

When it represents a boxed value, getUnboxedVal is meaningless, but when it represents an unboxed value, getBoxedVal tells us its type.

Constructors

Val

Fields

getUnboxedVal :: !UVal

getBoxedVal :: !BVal

Bundled Patterns

pattern CharVal :: Char -> Val

pattern NatVal :: Word64 -> Val

pattern DoubleVal :: Double -> Val

pattern IntVal :: Int -> Val

pattern BoolVal :: Bool -> Val

pattern UnboxedVal :: Int -> UnboxedTypeTag -> Val

pattern BoxedVal :: Closure -> Val
Matches a Val which is known to be boxed, and returns the closure portion.

emptyVal :: Val

A nulled out value you can use when filling empty arrays, etc.

falseVal :: Val

trueVal :: Val

boxedVal :: BVal -> Val

Lift a boxed val into an Val

type USeq = Seq Val

Implementation for Unison sequences.

traceK :: Reference -> K -> [(Reference, Int)]

frameDataSize :: K -> Int

data RuntimePanic

Constructors

Panic String (Maybe Val)

marshalUnwrapForeignIO :: (HasCallStack, BuiltinForeign a) => Closure -> IO a

unull :: USeg

bnull :: BSeg

nullSeg :: Seg

peekD :: Stack -> IO Double

peekOffD :: Stack -> Int -> IO Double

peekC :: Stack -> IO Char

peekOffC :: Stack -> Int -> IO Char

poke :: DebugCallStack => Stack -> Val -> IO ()

pokeD :: Stack -> Double -> IO ()

pokeOffD :: Stack -> Int -> Double -> IO ()

pokeC :: Stack -> Char -> IO ()

pokeOffC :: Stack -> Int -> Char -> IO ()

pokeBool :: DebugCallStack => Stack -> Bool -> IO ()

pokeTag :: DebugCallStack => Stack -> Int -> IO ()

Store an unboxed tag to later match on. Often used to indicate the constructor of a data type that's been unpacked onto the stack, or some tag we're about to branch on.

peekTag :: DebugCallStack => Stack -> IO Int

peekTagOff :: DebugCallStack => Stack -> Off -> IO Int

peekI :: DebugCallStack => Stack -> IO Int

peekOffI :: DebugCallStack => Stack -> Off -> IO Int

peekN :: Stack -> IO Word64

peekOffN :: Stack -> Int -> IO Word64

pokeN :: Stack -> Word64 -> IO ()

pokeOffN :: Stack -> Int -> Word64 -> IO ()

pokeI :: Stack -> Int -> IO ()

Note: This is for poking an unboxed value that has the UNISON type int, not just any unboxed data.

pokeOffI :: Stack -> Int -> Int -> IO ()

pokeByte :: Stack -> Word8 -> IO ()

peekBi :: BuiltinForeign b => Stack -> IO b

peekOffBi :: BuiltinForeign b => Stack -> Int -> IO b

pokeBi :: BuiltinForeign b => Stack -> b -> IO ()

pokeOffBi :: BuiltinForeign b => Stack -> Int -> b -> IO ()

peekBool :: Stack -> IO Bool

peekOffBool :: Stack -> Int -> IO Bool

peekOffS :: Stack -> Int -> IO USeq

pokeS :: Stack -> USeq -> IO ()

pokeOffS :: Stack -> Int -> USeq -> IO ()

frameView :: Stack -> IO ()

scount :: Seg -> Int

closureTermRefs :: Monoid m => (Reference -> m) -> Closure -> m

dumpAP :: Int -> Int -> Int -> Dump -> Int

dumpFP :: Int -> Int -> Dump -> Int

alloc :: IO Stack

peek :: DebugCallStack => Stack -> IO Val

upeek :: DebugCallStack => Stack -> IO UVal

bpeek :: DebugCallStack => Stack -> IO BVal

peekOff :: DebugCallStack => Stack -> Off -> IO Val

upeekOff :: DebugCallStack => Stack -> Off -> IO UVal

bpeekOff :: DebugCallStack => Stack -> Off -> IO BVal

bpoke :: DebugCallStack => Stack -> BVal -> IO ()

Store a boxed value. We don't bother nulling out the unboxed stack, it's extra work and there's nothing to garbage collect.

bpokeOff :: DebugCallStack => Stack -> Off -> BVal -> IO ()

pokeOff :: DebugCallStack => Stack -> Off -> Val -> IO ()

upokeT :: DebugCallStack => Stack -> UVal -> BVal -> IO ()

upokeOffT :: DebugCallStack => Stack -> Off -> UVal -> BVal -> IO ()

unsafePokeIasN :: DebugCallStack => Stack -> Int -> IO ()

Sometimes we get back an int from a foreign call which we want to use as a Nat. If we know it's positive and smaller than 2^63 then we can safely store the Int directly as a Nat without checks.

bump :: Stack -> IO Stack

bumpn :: Stack -> SZ -> IO Stack

grabSeg :: Stack -> SZ -> IO (Seg, Stack)

Eats up arguments

truncateSeg :: Stack -> SZ -> IO Stack

ensure :: Stack -> SZ -> IO Stack

duplicate :: Stack -> IO Stack

discardFrame :: Stack -> IO Stack

saveFrame :: Stack -> IO (Stack, SZ, SZ)

saveArgs :: Stack -> IO (Stack, SZ)

restoreFrame :: Stack -> SZ -> SZ -> IO Stack

prepareArgs :: Stack -> Args' -> IO Stack

acceptArgs :: Stack -> Int -> IO Stack

frameArgs :: Stack -> IO Stack

augSeg :: Augment -> Stack -> Seg -> Maybe Args' -> IO Seg

dumpSeg :: Stack -> Seg -> Dump -> IO Stack

adjustArgs :: Stack -> SZ -> IO Stack

fsize :: Stack -> SZ

asize :: Stack -> SZ

useg :: [Int] -> USeg

Converts a list of integers representing an unboxed segment back into the appropriate segment. Segments are stored backwards in the runtime, so this reverses the list.

bseg :: [Closure] -> BSeg

Converts a list of closures back to a boxed segment. Segments are stored backwards, so this reverses the contents.

segFromList :: SegList -> Seg

Converts from the list representation of a segment to the efficient stack form. Segments are stored backwards, so this reverses the contents.

traverseListToSeg :: (a -> IO Val) -> [a] -> IO Seg

traverseAccumSegToList :: (Val -> StateT s IO a) -> Seg -> StateT s IO [a]

Unboxed type tags

natTypeTag :: Closure

intTypeTag :: Closure

charTypeTag :: Closure

floatTypeTag :: Closure

hasNoAllocations :: Name -> Obligation

inflateMap :: Map Val Val -> Closure

deflateMap :: Closure -> Maybe (Map Val Val)

data HashAlgorithm where

Constructors

HashAlgorithm :: forall a. HashAlgorithm a => Reference -> a -> HashAlgorithm

data Tls

Constructors

Tls

Fields

socket :: Socket

context :: Context