Unison.Runtime.TypeTags

Documentation

class Tag t where

Methods

rawTag :: t -> Word64

newtype RTag

Constructors

RTag Word64

newtype CTag

Constructors

CTag Word16

newtype PackedTag

A combined tag, which is a packed representation of an RTag and a CTag

Constructors

PackedTag Word64

packTags :: RTag -> CTag -> PackedTag

unpackTags :: PackedTag -> (RTag, CTag)

maskTags :: PackedTag -> Word64

typeTag :: PackedTag -> Word64

anyTag :: PackedTag

floatTag :: PackedTag

natTag :: PackedTag

intTag :: PackedTag

charTag :: PackedTag

unitTag :: PackedTag

leftTag :: PackedTag

rightTag :: PackedTag

noneTag :: PackedTag

someTag :: PackedTag

falseTag :: PackedTag

trueTag :: PackedTag

pairTag :: PackedTag

failureTag :: PackedTag

noBufTag :: PackedTag

lineBufTag :: PackedTag

blockBufTag :: PackedTag

sizedBlockBufTag :: PackedTag

readModeTag :: PackedTag

writeModeTag :: PackedTag

appendModeTag :: PackedTag

readWriteModeTag :: PackedTag

seekAbsoluteTag :: PackedTag

seekRelativeTag :: PackedTag

seekEndTag :: PackedTag

exceptionTag :: Word64

exceptionRaiseTag :: PackedTag

stdInTag :: PackedTag

stdOutTag :: PackedTag

stdErrTag :: PackedTag

pureEffectTag :: PackedTag

A tag we use to represent the pure effect case.

seqViewElemTag :: PackedTag

seqViewEmptyTag :: PackedTag

mapTipTag :: PackedTag

mapBinTag :: PackedTag

setWrapTag :: PackedTag

jsonNullTag :: PackedTag

jsonBoolTag :: PackedTag

jsonObjTag :: PackedTag

jsonNumTag :: PackedTag

jsonTextTag :: PackedTag

jsonArrTag :: PackedTag

jsonParseErrorTag :: PackedTag

avroNullTag :: PackedTag

avroRecordTag :: PackedTag

avroBytesTag :: PackedTag

avroFixedTag :: PackedTag

avroArrayTag :: PackedTag

avroMapTag :: PackedTag

avroUnionTag :: PackedTag

avroEnumTag :: PackedTag

avroStringTag :: PackedTag

avroIntTag :: PackedTag

avroLongTag :: PackedTag

avroFloatTag :: PackedTag

avroDoubleTag :: PackedTag

avroBooleanTag :: PackedTag

avroReadSchemaNullTag :: PackedTag

avroReadSchemaBooleanTag :: PackedTag

avroReadSchemaIntTag :: PackedTag

avroReadSchemaLongTag :: PackedTag

avroReadSchemaFloatTag :: PackedTag

avroReadSchemaDoubleTag :: PackedTag

avroReadSchemaBytesTag :: PackedTag

avroReadSchemaStringTag :: PackedTag

avroReadSchemaArrayTag :: PackedTag

avroReadSchemaMapTag :: PackedTag

avroReadSchemaRecordTag :: PackedTag

avroReadSchemaEnumTag :: PackedTag

avroReadSchemaUnionTag :: PackedTag

avroReadSchemaFixedTag :: PackedTag

avroReadSchemaFreeUnionTag :: PackedTag

avroReadSchemaNamedTypeTag :: PackedTag

avroTypeNameTag :: PackedTag

avroLogicalIntDateTag :: PackedTag

avroLogicalIntTimeTag :: PackedTag

avroLogicalIntDecimalTag :: PackedTag

avroReadFloatFromInt32Tag :: PackedTag

avroReadFloatFromInt64Tag :: PackedTag

avroReadFloatTag :: PackedTag

avroLogicalLongTimeMicrosTag :: PackedTag

avroLogicalLongTimestampMillisTag :: PackedTag

avroLogicalLongTimestampMicrosTag :: PackedTag

avroLogicalLongLocalTimestampMillisTag :: PackedTag

avroLogicalLongLocalTimestampMicrosTag :: PackedTag

avroLogicalLongDecimalTag :: PackedTag

avroLogicalFixedDurationTag :: PackedTag

avroLogicalFixedDecimalTag :: PackedTag

avroLogicalStringUUIDTag :: PackedTag

avroReadDoubleFromInt32Tag :: PackedTag

avroReadDoubleFromInt64Tag :: PackedTag

avroReadDoubleFromFloatTag :: PackedTag

avroReadDoubleTag :: PackedTag

avroDecimalTag :: PackedTag

avroReadLongInt32Tag :: PackedTag

avroReadLongTag :: PackedTag

avroLogicalBytesDecimalTag :: PackedTag

avroLogicalStringUuidTag :: PackedTag

avroReadFieldTag :: PackedTag

avroFieldTag :: PackedTag

avroFieldStatusAsIsTag :: PackedTag

avroFieldStatusDefaultedTag :: PackedTag

avroFieldStatusIgnoredTag :: PackedTag

avroOrderAscendingTag :: PackedTag

avroOrderDescendingTag :: PackedTag

avroOrderIgnoreTag :: PackedTag

avroDefaultValueIntTag :: PackedTag

avroDefaultValueLongTag :: PackedTag

avroDefaultValueFloatTag :: PackedTag

avroDefaultValueDoubleTag :: PackedTag

avroDefaultValueBytesTag :: PackedTag

avroDefaultValueStringTag :: PackedTag

avroDefaultValueArrayTag :: PackedTag

avroDefaultValueMapTag :: PackedTag

avroDefaultValueRecordTag :: PackedTag

avroDefaultValueUnionTag :: PackedTag

avroDefaultValueFixedTag :: PackedTag

avroDefaultValueEnumTag :: PackedTag

avroDefaultValueNullTag :: PackedTag

avroDefaultValueBooleanTag :: PackedTag

avroSchemaNullTag :: PackedTag

avroSchemaBooleanTag :: PackedTag

avroSchemaIntTag :: PackedTag

avroSchemaLongTag :: PackedTag

avroSchemaFloatTag :: PackedTag

avroSchemaDoubleTag :: PackedTag

avroSchemaBytesTag :: PackedTag

avroSchemaStringTag :: PackedTag

avroSchemaArrayTag :: PackedTag

avroSchemaMapTag :: PackedTag

avroSchemaNamedTypeTag :: PackedTag

avroSchemaRecordTag :: PackedTag

avroSchemaEnumTag :: PackedTag

avroSchemaUnionTag :: PackedTag

avroSchemaFixedTag :: PackedTag

avroReadRecordTag :: PackedTag

avroFixedTypeTag :: PackedTag

avroRecordTypeTag :: PackedTag

avroEnumTypeTag :: PackedTag