Unison.Runtime.ANF

Documentation

minimizeCyclesOrCrash :: Var v => Term v a -> Term v a

pattern TVar :: Var v => v -> Term ANormalF v

pattern TLit :: Var v => Lit -> Term ANormalF v

pattern TBLit :: Var v => Lit -> Term ANormalF v

pattern TApp :: Var v => Func v -> [v] -> Term ANormalF v

pattern TApv :: Var v => v -> [v] -> Term ANormalF v

pattern TCom :: Var v => Reference -> [v] -> Term ANormalF v

pattern TCon :: Var v => Reference -> CTag -> [v] -> Term ANormalF v

pattern TKon :: Var v => v -> [v] -> Term ANormalF v

pattern TReq :: Var v => Reference -> CTag -> [v] -> Term ANormalF v

pattern TPrm :: Var v => POp -> [v] -> Term ANormalF v

pattern TFOp :: Var v => FOp -> [v] -> Term ANormalF v

pattern THnd :: Var v => [Reference] -> v -> Term ANormalF v -> Term ANormalF v

pattern TLet :: Var v => Direction Word16 -> v -> Mem -> Term ANormalF v -> Term ANormalF v -> Term ANormalF v

pattern TLetD :: Var v => v -> Mem -> Term ANormalF v -> Term ANormalF v -> Term ANormalF v

pattern TFrc :: Var v => v -> Term ANormalF v

pattern TLets :: Var v => Direction Word16 -> [v] -> [Mem] -> Term ANormalF v -> Term ANormalF v -> Term ANormalF v

pattern TName :: Var v => v -> Either Reference v -> [v] -> Term ANormalF v -> Term ANormalF v

pattern TBind :: Var v => Cte v -> ANormal v -> ANormal v

pattern TBinds :: Var v => [Cte v] -> ANormal v -> ANormal v

pattern TShift :: Var v => Reference -> v -> Term ANormalF v -> Term ANormalF v

pattern TMatch :: Var v => v -> Branched (Term ANormalF v) -> Term ANormalF v

data CompileExn

Constructors

CE CallStack (Pretty ColorText)

internalBug :: HasCallStack => String -> a

data Mem

Constructors

UN

BX

data Lit

Constructors

I Int64

N Word64

F Double

T Text

C Char

LM Referent

LY Reference

data Cacheability

Whether the evaluation of a given definition is cacheable or not. i.e. it's a top-level pure value.

Constructors

Cacheable

Uncacheable

data Direction a

Constructors

Indirect a

Direct

data SuperNormal v

Constructors

Lambda

Fields

conventions :: [Mem]

bound :: ANormal v

data SuperGroup v

Constructors

Rec

Fields

group :: [(v, SuperNormal v)]

entry :: SuperNormal v

data POp

Constructors

ADDI

SUBI

MULI

DIVI

SGNI

NEGI

MODI

POWI

SHLI

SHRI

INCI

DECI

LEQI

EQLI

ADDN

SUBN

MULN

DIVN

MODN

TZRO

LZRO

POPC

POWN

SHLN

SHRN

ANDN

IORN

XORN

COMN

INCN

DECN

LEQN

EQLN

ADDF

SUBF

MULF

DIVF

MINF

MAXF

LEQF

EQLF

POWF

EXPF

SQRT

LOGF

LOGB

ABSF

CEIL

FLOR

TRNF

RNDF

COSF

ACOS

COSH

ACSH

SINF

ASIN

SINH

ASNH

TANF

ATAN

TANH

ATNH

ATN2

CATT

TAKT

DRPT

SIZT

IXOT

UCNS

USNC

EQLT

LEQT

PAKT

UPKT

CATS

TAKS

DRPS

SIZS

CONS

SNOC

IDXS

BLDS

VWLS

VWRS

SPLL

SPLR

PAKB

UPKB

TAKB

DRPB

IXOB

IDXB

SIZB

FLTB

CATB

ITOF

NTOF

ITOT

NTOT

TTOI

TTON

TTOF

FTOT

FORK

EQLU

CMPU

EROR

MISS

CACH

LKUP

LOAD

CVLD

SDBX

VALU

TLTT

PRNT

INFO

TRCE

DBTX

ATOM

TFRC

SDBL

SDBV

type FOp = Word64

close :: (Var v, Monoid a) => Set v -> Term v a -> Term v a

saturate :: (Var v, Monoid a) => Map ConstructorReference Int -> Term v a -> Term v a

float :: Var v => Monoid a => Map v Reference -> Term v a -> (Term v a, Map Reference Reference, [(Reference, Term v a)], [(Reference, Term v a)])

floatGroup :: Var v => Monoid a => Map v Reference -> [(v, Term v a)] -> ([(v, Id)], [(Reference, Term v a)], [(Reference, Term v a)])

lamLift :: Var v => Monoid a => Map v Reference -> Term v a -> (Term v a, Map Reference Reference, [(Reference, Term v a)], [(Reference, Term v a)])

lamLiftGroup :: Var v => Monoid a => Map v Reference -> [(v, Term v a)] -> ([(v, Id)], [(Reference, Term v a)], [(Reference, Term v a)])

litRef :: Lit -> Reference

inlineAlias :: Var v => Monoid a => Term v a -> Term v a

addDefaultCases :: Var v => Monoid a => Text -> Term v a -> Term v a

data ANormalF v e

Constructors

ALet (Direction Word16) [Mem] e e

AName (Either Reference v) [v] e

ALit Lit

ABLit Lit

AMatch v (Branched e)

AShift Reference e

AHnd [Reference] v e

AApp (Func v) [v]

AFrc v

AVar v

Bundled Patterns

pattern AApv :: v -> [v] -> ANormalF v e

pattern ACom :: Reference -> [v] -> ANormalF v e

pattern ACon :: Reference -> CTag -> [v] -> ANormalF v e

pattern AKon :: v -> [v] -> ANormalF v e

pattern AReq :: Reference -> CTag -> [v] -> ANormalF v e

pattern APrm :: POp -> [v] -> ANormalF v e

pattern AFOp :: FOp -> [v] -> ANormalF v e

type ANormal = Term ANormalF

data RTag

data CTag

class Tag t where

Methods

rawTag :: t -> Word64

data GroupRef

Constructors

GR Reference Word64

data Code

Constructors

CodeRep (SuperGroup Symbol) Cacheability

type UBValue = Either Word64 Value

A value which is either unboxed or boxed.

type ValList = [UBValue]

A list of either unboxed or boxed values. Each slot is one of unboxed or boxed but not both.

data Value

Constructors

Partial GroupRef ValList

Data Reference Word64 ValList

Cont ValList Cont

BLit BLit

data Cont

Constructors

KE

Mark Word64 [Reference] (Map Reference Value) Cont

Push Word64 Word64 GroupRef Cont

data BLit

Constructors

Text Text

List (Seq Value)

TmLink Referent

TyLink Reference

Bytes Bytes

Quote Value

Code Code

BArr ByteArray

Pos Word64

Neg Word64

Char Char

Float Double

Arr (Array Value)

packTags :: RTag -> CTag -> Word64

unpackTags :: Word64 -> (RTag, CTag)

maskTags :: Word64 -> Word64

type ANFM v = ReaderT (Set v) (State (Word64, Word16, [(v, SuperNormal v)]))

data Branched e

Constructors

MatchIntegral (EnumMap Word64 e) (Maybe e)

MatchText (Map Text e) (Maybe e)

MatchRequest (Map Reference (EnumMap CTag ([Mem], e))) e

MatchEmpty

MatchData Reference (EnumMap CTag ([Mem], e)) (Maybe e)

MatchSum (EnumMap Word64 ([Mem], e))

MatchNumeric Reference (EnumMap Word64 e) (Maybe e)

Bundled Patterns

pattern MatchDataCover :: Reference -> EnumMap CTag ([Mem], e) -> Branched e

data Func v

Constructors

FVar v

FComb !Reference

FCont v

FCon !Reference !CTag

FReq !Reference !CTag

FPrim (Either POp FOp)

data SGEqv v

Constructors

NumDefns (SuperGroup v) (SuperGroup v)

DefnConventions (SuperNormal v) (SuperNormal v)

Subterms (ANormal v) (ANormal v)

equivocate :: Var v => SuperGroup v -> SuperGroup v -> Either (SGEqv v) ()

superNormalize :: Var v => Term v a -> SuperGroup v

anfTerm :: Var v => Term v a -> ANFM v (DNormal v)

codeGroup :: Code -> SuperGroup Symbol

valueTermLinks :: Value -> [Reference]

valueLinks :: Monoid a => (Bool -> Reference -> a) -> Value -> a

groupTermLinks :: Var v => SuperGroup v -> [Reference]

foldGroup :: Monoid m => (SuperGroup Symbol -> m) -> Code -> m

foldGroupLinks :: (Monoid r, Var v) => (Bool -> Reference -> r) -> SuperGroup v -> r

overGroup :: (SuperGroup Symbol -> SuperGroup Symbol) -> Code -> Code

overGroupLinks :: Var v => (Bool -> Reference -> Reference) -> SuperGroup v -> SuperGroup v

traverseGroup :: Applicative f => (SuperGroup Symbol -> f (SuperGroup Symbol)) -> Code -> f Code

traverseGroupLinks :: (Applicative f, Var v) => (Bool -> Reference -> f Reference) -> SuperGroup v -> f (SuperGroup v)

normalLinks :: (Applicative f, Var v) => (Bool -> Reference -> f Reference) -> SuperNormal v -> f (SuperNormal v)

prettyGroup :: Var v => String -> SuperGroup v -> ShowS

prettySuperNormal :: Var v => Int -> SuperNormal v -> ShowS

prettyANF :: Var v => Bool -> Int -> ANormal v -> ShowS