Unison.Runtime.ANF.Serialize

Documentation

data Version

Constructors

Transfer Word32

Hash Word32

data TmTag

Constructors

VarT

ForceT

AppT

HandleT

ShiftT

MatchT

LitT

NameRefT

NameVarT

LetDirT

LetIndT

BxLitT

data FnTag

Constructors

FVarT

FCombT

FContT

FConT

FReqT

FPrimT

FForeignT

data MtTag

Constructors

MIntT

MTextT

MReqT

MEmptyT

MDataT

MSumT

MNumT

data LtTag

Constructors

IT

NT

FT

TT

CT

LMT

LYT

data BLTag

Constructors

TextT

ListT

TmLinkT

TyLinkT

BytesT

QuoteT

CodeT

BArrT

PosT

NegT

CharT

FloatT

ArrT

CachedCodeT

data VaTag

Constructors

PartialT

DataT

ContT

BLitT

data CoTag

Constructors

KET

MarkT

PushT

index :: Eq v => [v] -> v -> Maybe Word64

deindex :: HasCallStack => [v] -> Word64 -> v

pushCtx :: [v] -> [v] -> [v]

putIndex :: MonadPut m => Word64 -> m ()

getIndex :: MonadGet m => m Word64

putVar :: MonadPut m => Eq v => [v] -> v -> m ()

getVar :: MonadGet m => [v] -> m v

putArgs :: MonadPut m => Eq v => [v] -> [v] -> m ()

getArgs :: MonadGet m => [v] -> m [v]

putCCs :: MonadPut m => [Mem] -> m ()

getCCs :: MonadGet m => m [Mem]

putGroup :: MonadPut m => Var v => Map Reference Word64 -> EnumMap FOp Text -> SuperGroup v -> m ()

getGroup :: MonadGet m => Var v => m (SuperGroup v)

putCode :: MonadPut m => EnumMap FOp Text -> Code -> m ()

getCode :: MonadGet m => Word32 -> m Code

putCacheability :: MonadPut m => Cacheability -> m ()

getCacheability :: MonadGet m => m Cacheability

putComb :: MonadPut m => Var v => Map Reference Word64 -> EnumMap FOp Text -> [v] -> SuperNormal v -> m ()

getFresh :: Var v => Word64 -> v

getComb :: MonadGet m => Var v => [v] -> Word64 -> m (SuperNormal v)

putNormal :: MonadPut m => Var v => Map Reference Word64 -> EnumMap FOp Text -> [v] -> ANormal v -> m ()

getNormal :: MonadGet m => Var v => [v] -> Word64 -> m (ANormal v)

putFunc :: MonadPut m => Var v => Map Reference Word64 -> EnumMap FOp Text -> [v] -> Func v -> m ()

getFunc :: MonadGet m => Var v => [v] -> m (Func v)

putPOp :: MonadPut m => POp -> m ()

getPOp :: MonadGet m => m POp

pOpCode :: POp -> Word16

pOpAssoc :: [(POp, Word16)]

pop2word :: Map POp Word16

word2pop :: Map Word16 POp

putLit :: MonadPut m => Lit -> m ()

getLit :: MonadGet m => m Lit

putBLit :: MonadPut m => Version -> BLit -> m ()

getBLit :: MonadGet m => Version -> m BLit

putRefs :: MonadPut m => [Reference] -> m ()

getRefs :: MonadGet m => m [Reference]

putBranches :: MonadPut m => Var v => Map Reference Word64 -> EnumMap FOp Text -> [v] -> Branched (ANormal v) -> m ()

getBranches :: MonadGet m => Var v => [v] -> Word64 -> m (Branched (ANormal v))

putCase :: MonadPut m => Var v => Map Reference Word64 -> EnumMap FOp Text -> [v] -> ([Mem], ANormal v) -> m ()

getCase :: MonadGet m => Var v => [v] -> Word64 -> m ([Mem], ANormal v)

putCTag :: MonadPut m => CTag -> m ()

getCTag :: MonadGet m => m CTag

putGroupRef :: MonadPut m => GroupRef -> m ()

getGroupRef :: MonadGet m => m GroupRef

putValue :: MonadPut m => Version -> Value -> m ()

putUBValue :: MonadPut m => Version -> UBValue -> m ()

getValue :: MonadGet m => Version -> m Value

putCont :: MonadPut m => Version -> Cont -> m ()

getCont :: MonadGet m => Version -> m Cont

deserializeCode :: ByteString -> Either String Code

serializeCode :: EnumMap FOp Text -> Code -> ByteString

serializeGroupForRehash :: Var v => EnumMap FOp Text -> Reference -> SuperGroup v -> ByteString

Serializes a SuperGroup for rehashing.

Expected as arguments are some code, and the Reference that refers to it. In particular, if the code refers to itself by reference, or if the code is part of a mututally-recursive set of definitions (which have a common hash), the reference used as part of that (mutual) recursion must be supplied.

Using that reference, we find all references in the code to that connected component. In the resulting byte string, those references are instead replaced by positions in a listing of the connected component. This means that the byte string is independent of the hash used for the self reference. Only the order matters (which is determined by the Reference). Then the bytes can be re-hashed to establish a new hash for the connected component. This operation should be idempotent as long as the indexing is preserved.

Supplying a Builtin reference is not supported. Such code shouldn't be subject to rehashing.

getVersionedValue :: MonadGet m => m Value

deserializeValue :: ByteString -> Either String Value

serializeValue :: Value -> ByteString

serializeValueForHash :: Value -> ByteString

valueVersion :: Word32

codeVersion :: Word32