Unison.Runtime.Interface

Documentation

startRuntime :: Bool -> RuntimeHost -> Text -> IO (Runtime Symbol)

withRuntime :: MonadUnliftIO m => Bool -> RuntimeHost -> Text -> (Runtime Symbol -> m a) -> m a

startNativeRuntime :: Text -> FilePath -> IO (Runtime Symbol)

standalone :: CCache -> Word64 -> IO StoredCache

runStandalone :: StoredCache -> CombIx -> IO (Either (Pretty ColorText) ())

data StoredCache

A version of the Code Cache designed to be serialized to disk as standalone bytecode.

decodeStandalone :: ByteString -> Either String (Text, Text, CombIx, StoredCache)

data RuntimeHost

Whether the runtime is hosted within a persistent session or as a one-off process. This affects the amount of clean-up and book-keeping the runtime does.

Constructors

OneOff

Persistent

data Runtime v

Constructors

Runtime

Fields

terminate :: IO ()

evaluate :: CodeLookup v IO () -> PrettyPrintEnv -> Term v -> IO (Either Error ([Error], Term v))

compileTo :: CompileOpts -> CodeLookup v IO () -> PrettyPrintEnv -> Reference -> FilePath -> IO (Maybe Error)

mainType :: Type v Ann

ioTestTypes :: NESet (Type v Ann)