Unison.Runtime.MCode

Documentation

data Args'

Constructors

Arg1 !Int

Arg2 !Int !Int

ArgN !(PrimArray Int)

ArgR !Int !Int

data Args

Constructors

ZArgs

VArg1 !Int

VArg2 !Int !Int

VArgR !Int !Int

VArgN !(PrimArray Int)

VArgV !Int

data RefNums

Constructors

RN

Fields

dnum :: Reference -> Word64

cnum :: Reference -> Word64

anum :: Reference -> Maybe Int

data MLit

Constructors

MI !Int

MN !Word64

MC !Char

MD !Double

MT !Text

MM !Referent

MY !Reference

data GInstr comb

Constructors

Prim1 !Prim1 !Int

Prim2 !Prim2 !Int !Int

RefCAS !Int !Int !Int

ForeignCall !Bool !ForeignFunc !Args

SetDyn !Word64 !Int

Capture !Word64

Name !(GRef comb) !Args

Info !String

Pack !Reference !PackedTag !Args

Lit !MLit

Print !Int

Reset !(EnumSet Word64)

Fork !Int

Atomically !Int

Seq !Args

TryForce !Int

SandboxingFailure !Text

type Instr = GInstr CombIx

type RInstr val = GInstr (RComb val)

data GSection comb

Constructors

App !Bool !(GRef comb) !Args

Call !Bool !CombIx comb !Args

Jump !Int !Args

Match !Int !(GBranch comb)

Yield !Args

Ins !(GInstr comb) !(GSection comb)

Let !(GSection comb) !CombIx !Int !(GSection comb)

Die String

Exit

DMatch !(Maybe Reference) !Int !(GBranch comb)

NMatch !(Maybe Reference) !Int !(GBranch comb)

RMatch !Int !(GSection comb) !(EnumMap Word64 (GBranch comb))

Bundled Patterns

pattern MatchT :: Int -> GSection comb -> Map Text (GSection comb) -> GSection comb

pattern MatchW :: Int -> GSection comb -> EnumMap Word64 (GSection comb) -> GSection comb

type RSection val = GSection (RComb val)

type Section = GSection CombIx

data GComb val comb

Constructors

Comb !(GCombInfo comb)

CachedVal !Word64 !val

Bundled Patterns

pattern Lam :: Int -> Int -> GSection comb -> GComb val comb

data GCombInfo comb

Constructors

LamI !Int !Int !(GSection comb)

type Comb = GComb Void CombIx

newtype RComb val

The fixed point of a GComb where all references to a Comb are themselves Combs.

Constructors

RComb

Fields

unRComb :: GComb val (RComb val)

type RCombInfo val = GCombInfo (RComb val)

type GCombs val comb = EnumMap Word64 (GComb val comb)

Map of combinators, parameterized by comb reference type

type RCombs val = GCombs val (RComb val)

data CombIx

Constructors

CIx !Reference !Word64 !Word64

data GRef comb

Constructors

Stk !Int

Env !CombIx comb

Dyn !Word64

type RRef val = GRef (RComb val)

type Ref = GRef CombIx

A reference to a combinator, parameterized by comb

data Prim1

Constructors

DECI

DECN

INCI

INCN

NEGI

SGNI

LZRO

TZRO

COMN

COMI

POPC

ABSF

EXPF

LOGF

SQRT

COSF

ACOS

COSH

ACSH

SINF

ASIN

SINH

ASNH

TANF

ATAN

TANH

ATNH

ITOF

NTOF

CEIL

FLOR

TRNF

RNDF

TRNC

NOTB

SIZT

USNC

UCNS

ITOT

NTOT

FTOT

TTOI

TTON

TTOF

PAKT

UPKT

VWLS

VWRS

SIZS

PAKB

UPKB

SIZB

FLTB

MISS

CACH

LKUP

LOAD

CVLD

VALU

TLTT

DBTX

SDBL

REFN

REFR

RRFC

TIKR

data Prim2

Constructors

ADDI

ADDN

SUBI

SUBN

MULI

MULN

DIVI

DIVN

MODI

MODN

SHLI

SHLN

SHRI

SHRN

POWI

POWN

EQLI

EQLN

NEQI

NEQN

LEQI

LEQN

LESI

LESN

ANDN

ANDI

IORN

IORI

XORN

XORI

EQLF

NEQF

LEQF

LESF

ADDF

SUBF

MULF

DIVF

ATN2

POWF

LOGB

MAXF

MINF

CAST

DRPN

ANDB

IORB

EQLU

CMPU

LEQU

LESU

DRPT

CATT

TAKT

IXOT

EQLT

LEQT

LEST

DRPS

CATS

TAKS

CONS

SNOC

IDXS

SPLL

SPLR

TAKB

DRPB

IDXB

CATB

IXOB

THRO

TRCE

SDBX

SDBV

REFW

data GBranch comb

Constructors

Test1 !Word64 !(GSection comb) !(GSection comb)

Test2 !Word64 !(GSection comb) !Word64 !(GSection comb) !(GSection comb)

TestW !(GSection comb) !(EnumMap Word64 (GSection comb))

TestT !(GSection comb) !(Map Text (GSection comb))

type Branch = GBranch CombIx

type RBranch val = GBranch (RComb val)

emitCombs :: Var v => RefNums -> Reference -> Word64 -> SuperGroup v -> EnumMap Word64 Comb

emitComb :: Var v => RefNums -> Reference -> Word64 -> RCtx v -> (Word64, SuperNormal v) -> EnumMap Word64 Comb

resolveCombs :: Maybe (EnumMap Word64 (RCombs val)) -> EnumMap Word64 (GCombs val CombIx) -> EnumMap Word64 (RCombs val)

lazily replace all references to combinators with the combinators themselves, tying the knot recursively when necessary.

sanitizeCombsOfForeignFuncs :: Bool -> Set ForeignFunc -> EnumMap Word64 (EnumMap Word64 (GComb Void CombIx)) -> EnumMap Word64 (EnumMap Word64 (GComb Void CombIx))

If running in a sandboxed environment, replace all restricted foreign functions with an error.

absurdCombs :: EnumMap Word64 (EnumMap Word64 (GComb Void cix)) -> EnumMap Word64 (GCombs any cix)

emptyRNs :: RefNums

argsToLists :: Args -> [Int]

countArgs :: Args -> Int

combRef :: CombIx -> Reference

combDeps :: GComb val comb -> [Word64]

combTypes :: GComb any comb -> [Word64]

prettyCombs :: Word64 -> EnumMap Word64 Comb -> ShowS

prettyComb :: (Show val, Show comb) => Word64 -> Word64 -> GComb val comb -> ShowS