Unison.Runtime.MCode

Documentation

data Args'

Constructors

Arg1 !Int

Arg2 !Int !Int

ArgN !(PrimArray Int)

ArgR !Int !Int

data Args

Constructors

ZArgs

VArg1 !Int

VArg2 !Int !Int

VArgR !Int !Int

VArgN !(PrimArray Int)

VArgV !Int

data RefNums

Constructors

RN

Fields

dnum :: Reference -> Word64

cnum :: Reference -> Word64

data MLit

Constructors

MI !Int

MD !Double

MT !Text

MM !Referent

MY !Reference

data GInstr comb

Constructors

UPrim1 !UPrim1 !Int

UPrim2 !UPrim2 !Int !Int

BPrim1 !BPrim1 !Int

BPrim2 !BPrim2 !Int !Int

ForeignCall !Bool !Word64 !Args

SetDyn !Word64 !Int

Capture !Word64

Name !(GRef comb) !Args

Info !String

Pack !Reference !Word64 !Args

Lit !MLit

BLit !Reference !Word64 !MLit

Print !Int

Reset !(EnumSet Word64)

Fork !Int

Atomically !Int

Seq !Args

TryForce !Int

type Instr = GInstr CombIx

type RInstr clos = GInstr (RComb clos)

data GSection comb

Constructors

App !Bool !(GRef comb) !Args

Call !Bool !CombIx comb !Args

Jump !Int !Args

Match !Int !(GBranch comb)

Yield !Args

Ins !(GInstr comb) !(GSection comb)

Let !(GSection comb) !CombIx !Int !(GSection comb)

Die String

Exit

DMatch !(Maybe Reference) !Int !(GBranch comb)

NMatch !(Maybe Reference) !Int !(GBranch comb)

RMatch !Int !(GSection comb) !(EnumMap Word64 (GBranch comb))

Bundled Patterns

pattern MatchT :: Int -> GSection comb -> Map Text (GSection comb) -> GSection comb

pattern MatchW :: Int -> GSection comb -> EnumMap Word64 (GSection comb) -> GSection comb

type RSection clos = GSection (RComb clos)

type Section = GSection CombIx

data GComb clos comb

Constructors

Comb !(GCombInfo comb)

CachedClosure !Word64 !clos

Bundled Patterns

pattern Lam :: Int -> Int -> GSection comb -> GComb clos comb

data GCombInfo comb

Constructors

LamI !Int !Int !(GSection comb)

type Comb = GComb Void CombIx

newtype RComb clos

The fixed point of a GComb where all references to a Comb are themselves Combs.

Constructors

RComb

Fields

unRComb :: GComb clos (RComb clos)

type RCombInfo clos = GCombInfo (RComb clos)

type GCombs clos comb = EnumMap Word64 (GComb clos comb)

Map of combinators, parameterized by comb reference type

type RCombs clos = GCombs clos (RComb clos)

data CombIx

Constructors

CIx !Reference !Word64 !Word64

data GRef comb

Constructors

Stk !Int

Env !CombIx comb

Dyn !Word64

type RRef clos = GRef (RComb clos)

type Ref = GRef CombIx

A reference to a combinator, parameterized by comb

data UPrim1

Constructors

DECI

INCI

NEGI

SGNI

LZRO

TZRO

COMN

POPC

ABSF

EXPF

LOGF

SQRT

COSF

ACOS

COSH

ACSH

SINF

ASIN

SINH

ASNH

TANF

ATAN

TANH

ATNH

ITOF

NTOF

CEIL

FLOR

TRNF

RNDF

data UPrim2

Constructors

ADDI

SUBI

MULI

DIVI

MODI

DIVN

MODN

SHLI

SHRI

SHRN

POWI

EQLI

LEQI

LEQN

ANDN

IORN

XORN

EQLF

LEQF

ADDF

SUBF

MULF

DIVF

ATN2

POWF

LOGB

MAXF

MINF

data BPrim1

Constructors

SIZT

USNC

UCNS

ITOT

NTOT

FTOT

TTOI

TTON

TTOF

PAKT

UPKT

VWLS

VWRS

SIZS

PAKB

UPKB

SIZB

FLTB

MISS

CACH

LKUP

LOAD

CVLD

VALU

TLTT

DBTX

SDBL

data BPrim2

Constructors

EQLU

CMPU

DRPT

CATT

TAKT

IXOT

EQLT

LEQT

LEST

DRPS

CATS

TAKS

CONS

SNOC

IDXS

SPLL

SPLR

TAKB

DRPB

IDXB

CATB

IXOB

THRO

TRCE

SDBX

SDBV

data GBranch comb

Constructors

Test1 !Word64 !(GSection comb) !(GSection comb)

Test2 !Word64 !(GSection comb) !Word64 !(GSection comb) !(GSection comb)

TestW !(GSection comb) !(EnumMap Word64 (GSection comb))

TestT !(GSection comb) !(Map Text (GSection comb))

type Branch = GBranch CombIx

type RBranch clos = GBranch (RComb clos)

emitCombs :: Var v => RefNums -> Reference -> Word64 -> SuperGroup v -> EnumMap Word64 Comb

emitComb :: Var v => RefNums -> Reference -> Word64 -> RCtx v -> (Word64, SuperNormal v) -> EnumMap Word64 Comb

resolveCombs :: Maybe (EnumMap Word64 (RCombs clos)) -> EnumMap Word64 (GCombs clos CombIx) -> EnumMap Word64 (RCombs clos)

lazily replace all references to combinators with the combinators themselves, tying the knot recursively when necessary.

absurdCombs :: EnumMap Word64 (EnumMap Word64 (GComb Void cix)) -> EnumMap Word64 (GCombs any cix)

emptyRNs :: RefNums

argsToLists :: Args -> [Int]

countArgs :: Args -> Int

combRef :: CombIx -> Reference

combDeps :: GComb clos comb -> [Word64]

combTypes :: GComb any comb -> [Word64]

prettyCombs :: Word64 -> EnumMap Word64 Comb -> ShowS

prettyComb :: (Show clos, Show comb) => Word64 -> Word64 -> GComb clos comb -> ShowS