Unison.Runtime.Machine

Documentation

type ActiveThreads = Maybe (IORef (Set ThreadId))

A ref storing every currently active thread. This is helpful for cleaning up orphaned threads when the main process completes.

We track threads when running in a host process like UCM, otherwise, in one-off environments Nothing is used and we don't bother tracking forked threads since they'll be cleaned up automatically on process termination.

type Tag = Word64

type DEnv = EnumMap Word64 Closure

type MCombs = RCombs Closure

type Combs = GCombs Void CombIx

type MSection = RSection Closure

type MBranch = RBranch Closure

type MInstr = RInstr Closure

type MComb = RComb Closure

type MRef = RRef Closure

data Tracer

Constructors

NoTrace

MsgTrace String String String

SimpleTrace String

data CCache

Constructors

CCache

Fields

foreignFuncs :: EnumMap Word64 ForeignFunc

sandboxed :: Bool

tracer :: Bool -> Closure -> Tracer

srcCombs :: TVar (EnumMap Word64 Combs)

combs :: TVar (EnumMap Word64 MCombs)

combRefs :: TVar (EnumMap Word64 Reference)

cacheableCombs :: TVar (EnumSet Word64)

tagRefs :: TVar (EnumMap Word64 Reference)

freshTm :: TVar Word64

freshTy :: TVar Word64

intermed :: TVar (Map Reference (SuperGroup Symbol))

refTm :: TVar (Map Reference Word64)

refTy :: TVar (Map Reference Word64)

sandbox :: TVar (Map Reference (Set Reference))

refNumsTm :: CCache -> IO (Map Reference Word64)

refNumsTy :: CCache -> IO (Map Reference Word64)

refNumTm :: CCache -> Reference -> IO Word64

refNumTy :: CCache -> Reference -> IO Word64

refNumTy' :: CCache -> Reference -> IO (Maybe Word64)

baseCCache :: Bool -> IO CCache

info :: Show a => String -> a -> IO ()

infos :: String -> String -> IO ()

stk'info :: Stack -> IO ()

eval0 :: CCache -> ActiveThreads -> MSection -> IO ()

mCombClosure :: CombIx -> MComb -> Closure

topDEnv :: EnumMap Word64 MCombs -> Map Reference Word64 -> Map Reference Word64 -> (DEnv, K -> K)

apply0 :: Maybe (Stack -> IO ()) -> CCache -> ActiveThreads -> Word64 -> IO ()

apply1 :: (Stack -> IO ()) -> CCache -> ActiveThreads -> Closure -> IO ()

jump0 :: (Stack -> IO ()) -> CCache -> ActiveThreads -> Closure -> IO ()

unitValue :: Closure

lookupDenv :: Word64 -> DEnv -> Closure

buildLit :: Reference -> Word64 -> MLit -> Closure

exec :: CCache -> DEnv -> ActiveThreads -> Stack -> K -> Reference -> MInstr -> IO (DEnv, Stack, K)

Execute an instruction

encodeExn :: Stack -> Either SomeException () -> IO Stack

numValue :: Maybe Reference -> Closure -> IO Word64

eval :: CCache -> DEnv -> ActiveThreads -> Stack -> K -> Reference -> MSection -> IO ()

Evaluate a section

forkEval :: CCache -> ActiveThreads -> Closure -> IO ThreadId

nestEval :: CCache -> ActiveThreads -> (Closure -> IO ()) -> Closure -> IO ()

atomicEval :: CCache -> ActiveThreads -> (Closure -> IO ()) -> Closure -> IO ()

enter :: CCache -> DEnv -> ActiveThreads -> Stack -> K -> Bool -> Args -> MComb -> IO ()

name :: Stack -> Args -> Closure -> IO Stack

apply :: CCache -> DEnv -> ActiveThreads -> Stack -> K -> Bool -> Args -> Closure -> IO ()

jump :: CCache -> DEnv -> ActiveThreads -> Stack -> K -> Args -> Closure -> IO ()

repush :: CCache -> ActiveThreads -> Stack -> DEnv -> K -> K -> IO ()

moveArgs :: Stack -> Args -> IO Stack

closureArgs :: Stack -> Args -> IO [Closure]

buildData :: Stack -> Reference -> Tag -> Args -> IO Closure

TODO: Experiment: In cases where we need to check the boxed stack to see where the argument lives we can either fetch from both unboxed and boxed stacks, then check the boxed result; OR we can just fetch from the boxed stack and check the result, then conditionally fetch from the unboxed stack.

The former puts more work before the branch, which _may_ be better for cpu pipelining, but the latter avoids an unnecessary fetch from the unboxed stack in cases where all args are boxed.

dumpDataNoTag :: Maybe Reference -> Stack -> Closure -> IO (Word64, Stack)

closeArgs :: Augment -> Stack -> Seg -> Args -> IO Seg

peekForeign :: Stack -> Int -> IO a

uprim1 :: Stack -> UPrim1 -> Int -> IO Stack

uprim2 :: Stack -> UPrim2 -> Int -> Int -> IO Stack

bprim1 :: Stack -> BPrim1 -> Int -> IO Stack

bprim2 :: Stack -> BPrim2 -> Int -> Int -> IO Stack

yield :: CCache -> DEnv -> ActiveThreads -> Stack -> K -> IO ()

selectTextBranch :: Text -> MSection -> Map Text MSection -> MSection

selectBranch :: Tag -> MBranch -> MSection

splitCont :: DEnv -> Stack -> K -> Word64 -> IO (Closure, DEnv, Stack, K)

discardCont :: DEnv -> Stack -> K -> Word64 -> IO (DEnv, Stack, K)

resolve :: CCache -> DEnv -> Stack -> MRef -> IO Closure

unhandledErr :: String -> CCache -> Word64 -> IO a

rCombSection :: EnumMap Word64 MCombs -> CombIx -> MComb

resolveSection :: CCache -> Section -> IO MSection

dummyRef :: Reference

reserveIds :: Word64 -> TVar Word64 -> IO Word64

updateMap :: Semigroup s => s -> TVar s -> STM s

refLookup :: String -> Map Reference Word64 -> Reference -> Word64

decodeCacheArgument :: Seq Closure -> IO [(Reference, Code)]

decodeSandboxArgument :: Seq Closure -> IO [Reference]

encodeSandboxListResult :: [Reference] -> Seq Closure

encodeSandboxResult :: Either [Reference] [Reference] -> Closure

encodeLeft :: Closure -> Closure

encodeRight :: Closure -> Closure

addRefs :: TVar Word64 -> TVar (Map Reference Word64) -> TVar (EnumMap Word64 Reference) -> Set Reference -> STM (Map Reference Word64)

codeValidate :: [(Reference, SuperGroup Symbol)] -> CCache -> IO (Maybe (Failure Closure))

sandboxList :: CCache -> Referent -> IO [Reference]

checkSandboxing :: CCache -> [Reference] -> Closure -> IO Bool

checkValueSandboxing :: CCache -> [Reference] -> Value -> IO (Either [Reference] [Reference])

evaluateSTM :: a -> STM a

cacheAdd0 :: Set Reference -> [(Reference, Code)] -> [(Reference, Set Reference)] -> CCache -> IO ()

preEvalTopLevelConstants :: EnumMap Word64 (GCombs Closure CombIx) -> EnumMap Word64 (GCombs Closure CombIx) -> CCache -> IO ()

expandSandbox :: Map Reference (Set Reference) -> [(Reference, SuperGroup Symbol)] -> [(Reference, Set Reference)]

cacheAdd :: [(Reference, Code)] -> CCache -> IO [Reference]

reflectValue :: EnumMap Word64 Reference -> Closure -> IO Value

reifyValue :: CCache -> Value -> IO (Either [Reference] Closure)

reifyValue0 :: (EnumMap Word64 MCombs, Map Reference Word64, Map Reference Word64) -> Value -> IO Closure

doubleToInt :: Double -> Int

intToDouble :: Int -> Double

closureNum :: Closure -> Int

universalEq :: (Foreign -> Foreign -> Bool) -> Closure -> Closure -> Bool

arrayEq :: (Closure -> Closure -> Bool) -> Array Closure -> Array Closure -> Bool

compareAsFloat :: Int -> Int -> Ordering

compareAsNat :: Int -> Int -> Ordering

floatTag :: Word64

natTag :: Word64

intTag :: Word64

charTag :: Word64

unitTag :: Word64

leftTag :: Word64

rightTag :: Word64

universalCompare :: (Foreign -> Foreign -> Ordering) -> Closure -> Closure -> Ordering

arrayCmp :: (Closure -> Closure -> Ordering) -> Array Closure -> Array Closure -> Ordering