Unison.Runtime.Stack

Documentation

data K

Constructors

KE

CB Callback

Mark !Int !(EnumSet Word64) !(EnumMap Word64 Closure) !K

Push !Int !Int !CombIx !Int !(RSection Closure) !K

data GClosure comb

Constructors

GPAp !CombIx !(GCombInfo comb) !Seg

GEnum !Reference !Word64

GDataU1 !Reference !Word64 !Int

GDataU2 !Reference !Word64 !Int !Int

GDataB1 !Reference !Word64 !(GClosure comb)

GDataB2 !Reference !Word64 !(GClosure comb) !(GClosure comb)

GDataUB !Reference !Word64 !Int !(GClosure comb)

GDataBU !Reference !Word64 !(GClosure comb) !Int

GDataG !Reference !Word64 !Seg

GCaptured !K !Int !Seg

GForeign !Foreign

GBlackHole

newtype Closure

Constructors

Closure

Fields

unClosure :: GClosure (RComb Closure)

Bundled Patterns

pattern DataC :: Reference -> Word64 -> SegList -> Closure

pattern PApV :: CombIx -> RCombInfo Closure -> SegList -> Closure

pattern CapV :: K -> Int -> SegList -> Closure

pattern PAp :: CombIx -> GCombInfo (RComb Closure) -> Seg -> Closure

pattern Enum :: Reference -> Word64 -> Closure

pattern DataU1 :: Reference -> Word64 -> Int -> Closure

pattern DataU2 :: Reference -> Word64 -> Int -> Int -> Closure

pattern DataB1 :: Reference -> Word64 -> Closure -> Closure

pattern DataB2 :: Reference -> Word64 -> Closure -> Closure -> Closure

pattern DataUB :: Reference -> Word64 -> Int -> Closure -> Closure

pattern DataBU :: Reference -> Word64 -> Closure -> Int -> Closure

pattern DataG :: Reference -> Word64 -> Seg -> Closure

pattern Captured :: K -> Int -> Seg -> Closure

pattern Foreign :: Foreign -> Closure

pattern BlackHole :: Closure

type IxClosure = GClosure CombIx

newtype Callback

Constructors

Hook (Stack -> IO ())

data Augment

Constructors

I

K

C

data Dump

Constructors

A

F Int Int

S

data Stack

Constructors

Stack

Fields

ap :: !Int

fp :: !Int

sp :: !Int

ustk :: !(MutableByteArray (PrimState IO))

bstk :: !(MutableArray (PrimState IO) Closure)

type Off = Int

type SZ = Int

type FP = Int

type Seg = (USeg, BSeg)

type USeg = ByteArray

type BSeg = Array Closure

traceK :: Reference -> K -> [(Reference, Int)]

frameDataSize :: K -> Int

marshalToForeign :: HasCallStack => Closure -> Foreign

unull :: USeg

bnull :: BSeg

nullSeg :: Seg

peekD :: Stack -> IO Double

peekOffD :: Stack -> Int -> IO Double

pokeD :: Stack -> Double -> IO ()

pokeOffD :: Stack -> Int -> Double -> IO ()

peekN :: Stack -> IO Word64

peekOffN :: Stack -> Int -> IO Word64

pokeN :: Stack -> Word64 -> IO ()

pokeOffN :: Stack -> Int -> Word64 -> IO ()

peekBi :: BuiltinForeign b => Stack -> IO b

peekOffBi :: BuiltinForeign b => Stack -> Int -> IO b

pokeBi :: BuiltinForeign b => Stack -> b -> IO ()

pokeOffBi :: BuiltinForeign b => Stack -> Int -> b -> IO ()

peekOffS :: Stack -> Int -> IO (Seq Closure)

pokeS :: Stack -> Seq Closure -> IO ()

pokeOffS :: Stack -> Int -> Seq Closure -> IO ()

frameView :: Stack -> IO ()

scount :: Seg -> Int

closureTermRefs :: Monoid m => (Reference -> m) -> Closure -> m

dumpAP :: Int -> Int -> Int -> Dump -> Int

dumpFP :: Int -> Int -> Dump -> Int

alloc :: IO Stack

peek :: Stack -> IO Elem

upeek :: Stack -> IO UElem

bpeek :: Stack -> IO BElem

peekOff :: Stack -> Off -> IO Elem

upeekOff :: Stack -> Off -> IO UElem

bpeekOff :: Stack -> Off -> IO BElem

bpoke :: Stack -> BElem -> IO ()

Store a boxed value. We don't bother nulling out the unboxed stack, it's extra work and there's nothing to garbage collect.

bpokeOff :: Stack -> Off -> BElem -> IO ()

upoke :: Stack -> UElem -> IO ()

Store an unboxed value and null out the boxed stack at that location, both so we know there's no value there, and so garbage collection can clean up any value that was referenced there.

upokeOff :: Stack -> Off -> UElem -> IO ()

bump :: Stack -> IO Stack

bumpn :: Stack -> SZ -> IO Stack

grab :: Stack -> SZ -> IO (Seg, Stack)

Eats up arguments

ensure :: Stack -> SZ -> IO Stack

duplicate :: Stack -> IO Stack

discardFrame :: Stack -> IO Stack

saveFrame :: Stack -> IO (Stack, SZ, SZ)

saveArgs :: Stack -> IO (Stack, SZ)

restoreFrame :: Stack -> SZ -> SZ -> IO Stack

prepareArgs :: Stack -> Args' -> IO Stack

acceptArgs :: Stack -> Int -> IO Stack

frameArgs :: Stack -> IO Stack

augSeg :: Augment -> Stack -> Seg -> Maybe Args' -> IO Seg

dumpSeg :: Stack -> Seg -> Dump -> IO Stack

adjustArgs :: Stack -> SZ -> IO Stack

fsize :: Stack -> SZ

asize :: Stack -> SZ