Unison.Runtime.TypeTags

Documentation

class Tag t where

Methods

rawTag :: t -> Word64

newtype RTag

Constructors

RTag Word64

newtype CTag

Constructors

CTag Word16

newtype PackedTag

A combined tag, which is a packed representation of an RTag and a CTag

Constructors

PackedTag Word64

packTags :: RTag -> CTag -> PackedTag

unpackTags :: PackedTag -> (RTag, CTag)

maskTags :: PackedTag -> Word64

anyTag :: PackedTag

floatTag :: PackedTag

natTag :: PackedTag

intTag :: PackedTag

charTag :: PackedTag

unitTag :: PackedTag

leftTag :: PackedTag

rightTag :: PackedTag

noneTag :: PackedTag

someTag :: PackedTag

falseTag :: PackedTag

trueTag :: PackedTag

pairTag :: PackedTag

failureTag :: PackedTag

noBufTag :: PackedTag

lineBufTag :: PackedTag

blockBufTag :: PackedTag

sizedBlockBufTag :: PackedTag

readModeTag :: PackedTag

writeModeTag :: PackedTag

appendModeTag :: PackedTag

readWriteModeTag :: PackedTag

seekAbsoluteTag :: PackedTag

seekRelativeTag :: PackedTag

seekEndTag :: PackedTag

exceptionTag :: Word64

exceptionRaiseTag :: PackedTag

stdInTag :: PackedTag

stdOutTag :: PackedTag

stdErrTag :: PackedTag

pureEffectTag :: PackedTag

A tag we use to represent the pure effect case.

seqViewElemTag :: PackedTag

seqViewEmptyTag :: PackedTag